refactor(fullscreen): Use Teleport to add the full-screen function to the root of the body and componentize it.
