+++ docs/vite.config.js
@@ -8,6 +8,10 @@ export default defineConfig({
         __dirname,
         "../dist/Mermaid.vue"
       ),
+      "vitepress-plugin-mermaid/Mermaid.vue": path.join(
+        __dirname,
+        "../dist/DragMermaid.vue"
+      ),
     },
   },
   server: {
